+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file. This change
 
 ### Fixed
 - Consistent use of 'accept' in `planck.http` ([#837](https://github.com/planck-repl/planck/issues/837))
+- `planck.core/load-string` evaluates in current namespace ([#867](https://github.com/planck-repl/planck/issues/867))
 
 ## [2.20.0] - 2019-01-31
 ### Added